move `Array` grouping to stable ES
