+++ packages/core-js/actual/map/group-by.js
@@ -1,13 +1,5 @@
 'use strict';
-require('../../modules/es.map');
+var parent = require('../../stable/map/group-by');
 require('../../modules/esnext.map.group-by');
-var call = require('../../internals/function-call');
-var isCallable = require('../../internals/is-callable');
-var path = require('../../internals/path');
 
-var Map = path.Map;
-var mapGroupBy = Map.groupBy;
-
-module.exports = function groupBy(source, iterable, keyDerivative) {
-  return call(mapGroupBy, isCallable(this) ? this : Map, source, iterable, keyDerivative);
-};
+module.exports = parent;
