+++ packages/core-js/actual/object/group-by.js
@@ -1,7 +1,5 @@
 'use strict';
-require('../../modules/es.object.create');
+var parent = require('../../stable/object/group-by');
 require('../../modules/esnext.object.group-by');
 
-var path = require('../../internals/path');
-
-module.exports = path.Object.groupBy;
+module.exports = parent;
