+++ packages/core-js/es/map/group-by.js
@@ -1,0 +1,13 @@
+'use strict';
+require('../../modules/es.map');
+require('../../modules/es.map.group-by');
+var call = require('../../internals/function-call');
+var isCallable = require('../../internals/is-callable');
+var path = require('../../internals/path');
+
+var Map = path.Map;
+var mapGroupBy = Map.groupBy;
+
+module.exports = function groupBy(source, iterable, keyDerivative) {
+  return call(mapGroupBy, isCallable(this) ? this : Map, source, iterable, keyDerivative);
+};
