+++ packages/core-js/es/map/index.js
@@ -1,6 +1,7 @@
 'use strict';
 require('../../modules/es.array.iterator');
 require('../../modules/es.map');
+require('../../modules/es.map.group-by');
 require('../../modules/es.object.to-string');
 require('../../modules/es.string.iterator');
 var path = require('../../internals/path');
