+++ packages/core-js/es/object/group-by.js
@@ -1,0 +1,7 @@
+'use strict';
+require('../../modules/es.object.create');
+require('../../modules/es.object.group-by');
+
+var path = require('../../internals/path');
+
+module.exports = path.Object.groupBy;
