+++ packages/core-js/es/object/index.js
@@ -11,6 +11,7 @@ require('../../modules/es.object.get-own-property-descriptor');
 require('../../modules/es.object.get-own-property-descriptors');
 require('../../modules/es.object.get-own-property-names');
 require('../../modules/es.object.get-prototype-of');
+require('../../modules/es.object.group-by');
 require('../../modules/es.object.has-own');
 require('../../modules/es.object.is');
 require('../../modules/es.object.is-extensible');
