+++ packages/core-js/stable/map/group-by.js
@@ -1,0 +1,4 @@
+'use strict';
+var parent = require('../../es/map/group-by');
+
+module.exports = parent;
