+++ packages/core-js/stable/object/group-by.js
@@ -1,0 +1,4 @@
+'use strict';
+var parent = require('../../es/object/group-by');
+
+module.exports = parent;
