+++ packages/core-js/stage/4.js
@@ -2,6 +2,7 @@
 // TODO: Remove this entry from `core-js@4`
 require('../proposals/accessible-object-hasownproperty');
 require('../proposals/array-find-from-last');
+require('../proposals/array-grouping-v2');
 require('../proposals/change-array-by-copy-stage-4');
 // require('../proposals/error-cause');
 require('../proposals/global-this');
