+++ tests/entries/content.mjs
@@ -128,9 +128,9 @@ superset('core-js/full/url', /^web\.url(?:\.|$)/);
 superset('core-js/full/url-search-params', /^web\.url-search-params/);
 superset('core-js/full/weak-map', /^(?:es|esnext)\.weak-map/);
 superset('core-js/full/weak-set', /^(?:es|esnext)\.weak-set/);
-equal('core-js/proposals', /^(?:es\.(?:map|string\.at)|esnext\.|web\.url)/);
-equal('core-js/stage', /^(?:es\.(?:map|string\.at)|esnext\.|web\.url)/);
-equal('core-js/stage/pre', /^(?:es\.(?:map|string\.at)|esnext\.|web\.url)/);
+subset('core-js/proposals', /^(?:es\.(?:map|string\.at)|esnext\.|web\.url)/);
+subset('core-js/stage', /^(?:es\.(?:map|string\.at)|esnext\.|web\.url)/);
+subset('core-js/stage/pre', /^(?:es\.(?:map|string\.at)|esnext\.|web\.url)/);
 subset('core-js/stage/0', /^(?:es\.(?:map|string\.at)|esnext\.|web\.url)/);
 subset('core-js/stage/1', /^(?:es\.(?:map|string\.at)|esnext\.|web\.url)/);
 subset('core-js/stage/2', /^es\.string\.at|esnext\./);
