+++ tests/unit-global/es.map.group-by.js
