+++ tests/unit-global/es.object.group-by.js
