+++ tests/unit-pure/es.object.group-by.js
@@ -1,5 +1,5 @@
 import { createIterable } from '../helpers/helpers.js';
-import groupBy from 'core-js-pure/actual/object/group-by';
+import groupBy from 'core-js-pure/es/object/group-by';
 import getPrototypeOf from 'core-js-pure/es/object/get-prototype-of';
 import entries from 'core-js-pure/es/object/entries';
